renderer: change storage of discardMode from uint32_t to uint8_t

+++ src/render/pass/SurfacePassElement.hpp
@@ -43,16 +43,16 @@ class CSurfacePassElement : public IPassElement {
         bool      popup = false;
 
         // counts how many surfaces this pass has rendered
-        int      surfaceCounter = 0;
+        int     surfaceCounter = 0;
 
-        CBox     clipBox = {}; // scaled coordinates
+        CBox    clipBox = {}; // scaled coordinates
 
-        uint32_t discardMode    = DISCARD_OPAQUE;
-        float    discardOpacity = 0.f;
+        uint8_t discardMode    = DISCARD_OPAQUE;
+        float   discardOpacity = 0.f;
 
-        bool     useNearestNeighbor = false;
+        bool    useNearestNeighbor = false;
 
-        bool     flipEndFrame = false;
+        bool    flipEndFrame = false;
     };
 
     CSurfacePassElement(const SRenderData& data);
